version 11032014

+++ templates/ezwebhosting/vertex/compression/top_css_and_js_compression.php
@@ -1,15 +1,11 @@
 <?php 
 $content = ob_get_contents();
 ob_clean();
-
 include_once('class_css_and_js_compression.php');
-
 $excludeFilesArray  = $excludeFilesArrayTop;
-
 // js regax
 $regex = '/src=\"([^\"]*)\"/i';
 $jsfiles = getArray($content,$regex);
-
 foreach($jsfiles as $k=>$jsfile){
 	foreach($excludeFilesArray as $excludeFiles){
 		if( stristr( $jsfile , $excludeFiles )){
@@ -19,8 +15,6 @@ foreach($jsfiles as $k=>$jsfile){
 		}
 	}
 }
-
-
 $regex = '/href=\"([^\"]*)\"/i';
 $cssfiles = getArray($content,$regex);
 
@@ -31,12 +25,13 @@ foreach($cssfiles as $k=>$cssfile){
 			 unset($cssfiles[$k]);
 			 break;
 		}
+		if((strpos($cssfiles[$k],'.ico') > -1 ) || (strpos($cssfiles[$k],'googleapis') > -1 )){
+			$excludecssFilesArray[] = $cssfiles[$k];
+			unset($cssfiles[$k]);
+			break;
+		}
 	}
 }
-
-
-
-
 $arrayTemp = array();
 if(isset($usejscompression) && ($usejscompression=="yes")){
 	foreach($jsfiles as $jsfile){
@@ -55,9 +50,7 @@ if(isset($usejscompression) && ($usejscompression=="yes")){
 			}
 		}
 	}
-	
 }
-
 if(isset($usecsscompression) && ($usecsscompression=="yes")){
 	foreach($cssfiles as $cssfil){
 		if(!in_array($cssfil,$arrayTemp)) {
@@ -74,13 +67,13 @@ if(isset($usecsscompression) && ($usecsscompression=="yes")){
 		}
 	}
 }
+ob_end_clean();
+ob_start();
 echo $content = removeblankLinks($content);
 /*echo "<pre>";
 print_r($jsfiles);
 print_r($cssfiles);
 exit;*/
-
-
 $HTTP_HOST =  $_SERVER['HTTP_HOST'];
 $SCRIPT_NAME =  $_SERVER['SCRIPT_NAME'];
 $scripnameArray = explode("/",$SCRIPT_NAME);
@@ -95,60 +88,84 @@ $scriptpath = $scripname."/".$filepath;
 $writabledir=$filepath;
 $mybrowser = getBrowser();
 $name = 'top_compression_js.php';
-
-if(isset($usejscompression) && ($usejscompression=="yes")){
+$writefile = delete_old_md5s($writabledir,$name,$cachetime);
+if(isset($usejscompression) && ($usejscompression=="yes")  && $writefile){
 /* 	if(file_exists($writabledir.$name)){
 			delete_old_md5s($writabledir,$name,$cachetime);
-		 
-	
 	}else{*/
+			/*temp added */
+			/*foreach($jsfiles as $jsfile){
+				if(strpos($jsfile,'.js') > -1)
+				echo $jsfile."<br>";	
+			}*/
+			/* end temp added */
 			$js = '';
 			foreach($jsfiles as $jsfile){
 				$js .= JSMin::minify(file_get_contents($jsfile));
 			}
-			if(delete_old_md5s($writabledir,$name,$cachetime)){
+			if($writefile){
 				$fh = fopen($writabledir.$name, 'w');
-				$js = '<?php header("Content-type: text/javascript"); ?>' . $js . '<?php exit();?>' ;
+				$js = '<?php ob_clean();header("Content-type: application/javascript",true); ob_start();?>' . $js . '<?php ob_flush();exit();?>' ;
 				fwrite($fh, $js);
 				fclose($fh);
 			}
+			
+			/*echo "<pre>";
+				print_r($jsfiles);
+			echo "</pre>";*/
 			//file_put_contents($writabledir.$name,$js);
-			
-			
 	//}
 	echo "\n
 			<script type='text/javascript' src='$serverpath"."$name' > </script>" ;
+} else if(isset($usejscompression) && ($usejscompression=="yes")  && !$writefile){
+	echo "\n
+			<script type='text/javascript' src='$serverpath"."$name' > </script>" ;	
 }
-
 $name = 'top_compression_css.php';
-if(isset($usecsscompression) && ($usecsscompression=="yes")){ 
+$writefile = delete_old_md5s($writabledir,$name,$cachetime);
+if(isset($usecsscompression) && ($usecsscompression=="yes") && $writefile){ 
 /*	if(file_exists($writabledir.$name)){
 	delete_old_md5s($writabledir,$name,$cachetime);
-		 
-		
 	}else{*/
+	
+		/*temp added */
+		/*foreach($cssfiles as $cssfile){
+			if(strpos($cssfile,'.css') > -1)
+			echo $cssfile."<br>";	
+		}*/
+		/* end temp added */
+			
 		$css = '';
+		$favicon = '';
 		foreach($cssfiles as $cssfile){
+			if(strpos($cssfil,'.ico') > -1){
+				$favicon = $cssfil;
+				continue;
+			}
 			$css .= CssMin::minify(file_get_contents($cssfile));
 		}
 		$css = removeblankLinks($css);
-		if(delete_old_md5s($writabledir,$name,$cachetime)){
+		if($writefile){
 			$fh = fopen($writabledir.$name, 'w');
-			$css = '<?php header("Content-type: text/css"); ?>' . $css . '<?php exit();?>' ;
+			$css = '<?php ob_clean();header("Content-type: text/css",true); ob_start();?>' . $css . '<?php ob_flush();exit();?>' ;
 			$css = removepath($css);
 			fwrite($fh, $css);
 			fclose($fh);
 		}
-		//file_put_contents($writabledir.$name,$css);
 		
-	
+		/*echo "<pre>";
+			print_r($cssfiles);
+		echo "</pre>";*/
+		//file_put_contents($writabledir.$name,$css);
 	//}
-	
 		echo "\n
 		<link rel='stylesheet' href='$serverpath"."$name' type='text/css' />
 		" ;
+} else if(isset($usecsscompression) && ($usecsscompression=="yes") && !$writefile){
+	echo "\n
+		<link rel='stylesheet' href='$serverpath"."$name' type='text/css' />
+		" ;	
 }
-
 if(isset($excludejsFilesArray)){
 	if(is_array($excludejsFilesArray)){
 		foreach($excludejsFilesArray as $excludejsFiles){
@@ -159,9 +176,12 @@ if(isset($excludejsFilesArray)){
 if(isset($excludecssFilesArray)){
 	if(is_array($excludecssFilesArray)){
 		foreach($excludecssFilesArray as $excludecssFiles){
+			if(strpos($excludecssFiles,'.ico') > -1)
+				echo "\n<link type='image/x-icon' rel='shortcut icon' href='$excludecssFiles'>";
+			else
 				echo "\n<link rel='stylesheet' href='$excludecssFiles' type='text/css' />" ;
 		}
 	}
 }
- 
+
 ?>
